Move app to src/ and update typing
