+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -15,6 +15,6 @@ _Describe more what you did on changes._
 ### Checks
 
 If this is for a release:
- [ ] I have updated the version number in `pycov3/VERSION`
+ [ ] I have updated the version number in `src/pycov3/__init__.py`
  [ ] I have updated the README with any appropriate documentation
  [ ] I have updated `tests` with any appropriate tests
