+++ src/pycov3/Sequence.py
